+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add support for manageable oracle
+- add support for Silo V3
+
+### Updated
+- better validation
+
+### Removed
+- remove connected wallet as owner
 
 ## 0.8.0 - 2026-02-19
 ### Updated